logo_detector_app/requirements.txt
torch
transformers
Pillow
google-genai
streamlit
dotenv